Begin In-house exposure interface
Asks for in-house exposures for any opposite pairs of contagious/non-contagios persons.

When Person A has an exposure to Person B, appends the symmetric exposure of person B to Person A.

Doesn't change the calculator.

Behavior is kind of buggy.

Because of how the useState hook works differently from React States, I'm having trouble getting field states to stay matched with actual underlying data (try doing an edit, delete a field, hit cancel, and hit edit again. The field ought to be re-populated, but it is not.)

Also the symmetric exposure information is not deleted if you remove one person's exposure event.

Stopping here with it unfinshed, sorry :(( .