Abstracted the Date questions somewhat.
Using strings as CovidEvent keys, to make the cookie-cutter Date questions possible. Calculator appears to be unaffected.

Plus basic date text completion if the user adds no year or a 2-digit year.

Added onFocus and onBlur triggers to be used later, when we click a date to select.

No input validation yet, still :((